+++ scripts/prepublishOnly.ts
@@ -1,5 +1,9 @@
-import { readJsonSync, writeJsonSync } from 'fs-extra'
+import { outputFileSync, readJsonSync, writeJsonSync } from 'fs-extra'
 import path from 'path'
+
+type Exports = {
+  [key: string]: string | { types?: string; import: string; default: string }
+}
 
 generatePackageJson()
 
@@ -31,6 +35,34 @@ function generatePackageJson() {
     authors,
     keywords,
   } = tmpPackageJson
+
+  // Generate proxy packages for each export.
+  const files_ = [...files]
+  for (const [key, value] of Object.entries(exports_ as Exports)) {
+    if (typeof value === 'string') continue
+    if (key === '.') continue
+    if (!value.default || !value.import)
+      throw new Error('`default` and `import` are required.')
+
+    outputFileSync(
+      `${key}/package.json`,
+      `{
+  ${Object.entries(value)
+    .map(([k, v]) => {
+      const key = (() => {
+        if (k === 'import') return 'module'
+        if (k === 'default') return 'main'
+        if (k === 'types') return 'types'
+        throw new Error('Invalid key')
+      })()
+      return `"${key}": "${v.replace('./', '../')}"`
+    })
+    .join(',\n  ')}
+}`,
+    )
+    files_.push(key.replace('./', ''))
+  }
+
   writeJsonSync(
     packageJsonPath,
     {
@@ -39,7 +71,7 @@ function generatePackageJson() {
       dependencies,
       peerDependencies,
       version,
-      files,
+      files: files_,
       exports: exports_,
       // type,
       main,