+++ src/types/contract.ts
@@ -20,7 +20,7 @@ import type {
 
 import type { Hex, LogTopic } from './misc.js'
 import type { TransactionRequest } from './transaction.js'
-import type { Filter, MaybeRequired, NoUndefined, Prettify } from './utils.js'
+import type { Filter, NoUndefined } from './utils.js'
 
 export type AbiItem = Abi[number]
 
@@ -236,7 +236,10 @@ export type GetEventArgsFromTopics<
   TAbiEvent extends AbiEvent & { type: 'event' } = TAbi extends Abi
     ? ExtractAbiEvent<TAbi, TEventName>
     : AbiEvent & { type: 'event' },
-  TArgs = AbiEventTopicsToPrimitiveTypes<TAbiEvent['inputs'], TTopics, TData>,
+  TArgs = AbiEventParametersToPrimitiveTypes<
+    TAbiEvent['inputs'],
+    { EnableUnion: false; IndexedOnly: false }
+  >,
 > = TTopics extends readonly []
   ? TData extends undefined
     ? { args?: never }
@@ -249,12 +252,10 @@ export type GetEventArgsFromTopics<
 type EventParameterOptions = {
   EnableUnion?: boolean
   IndexedOnly?: boolean
-  Required?: boolean
 }
 type DefaultEventParameterOptions = {
   EnableUnion: true
   IndexedOnly: true
-  Required: false
 }
 
 type HashedEventTypes = 'bytes' | 'string' | 'tuple' | `${string}[${string}]`
@@ -326,13 +327,11 @@ export type AbiEventParametersToPrimitiveTypes<
               >
             },
           ]
-        | (Options['Required'] extends true
-            ? never
-            : // Distribute over tuple to represent optional parameters
-            Filtered extends readonly [
-                ...infer Head extends readonly AbiParameter[],
-                infer _,
-              ]
+        // Distribute over tuple to represent optional parameters
+        | (Filtered extends readonly [
+            ...infer Head extends readonly AbiParameter[],
+            infer _,
+          ]
             ? AbiEventParametersToPrimitiveTypes<
                 readonly [...{ [K in keyof Head]: Omit<Head[K], 'name'> }],
                 Options
@@ -347,81 +346,6 @@ export type AbiEventParametersToPrimitiveTypes<
             ? Name
             : never]?: AbiEventParameterToPrimitiveType<Parameter, Options>
       } extends infer Mapped
-    ? MaybeRequired<
-        Mapped,
-        Options['Required'] extends boolean ? Options['Required'] : false
-      >
+    ? Mapped
     : never
   : never
-
-/**
- * @internal
- */
-// TODO(@tmm): Simplify this type
-export type AbiEventTopicsToPrimitiveTypes<
-  TAbiParameters extends readonly AbiParameter[],
-  TTopics extends LogTopic[] | undefined = undefined,
-  TData extends Hex | undefined = undefined,
-  TBase = TAbiParameters[0] extends { name: string } ? {} : [],
-> = Prettify<
-  TAbiParameters extends readonly [
-    infer Head extends AbiParameter,
-    ...infer Tail,
-  ]
-    ? TTopics extends readonly [infer TopicHead, ...infer TopicTail]
-      ? Head extends { indexed: true }
-        ? Head extends { name: infer Name extends string }
-          ? {
-              [_ in Name]: TopicHead extends LogTopic
-                ? AbiEventTopicToPrimitiveType<Head, TopicHead>
-                : never
-            } & (Tail extends readonly []
-              ? {}
-              : Tail extends readonly AbiParameter[]
-              ? TopicTail extends LogTopic[]
-                ? AbiEventTopicsToPrimitiveTypes<Tail, TopicTail, TData>
-                : {}
-              : {})
-          : [
-              TopicHead extends LogTopic
-                ? AbiEventTopicToPrimitiveType<Head, TopicHead>
-                : never,
-              ...(Tail extends readonly []
-                ? []
-                : Tail extends readonly AbiParameter[]
-                ? TopicTail extends LogTopic[]
-                  ? AbiEventTopicsToPrimitiveTypes<Tail, TopicTail, TData>
-                  : []
-                : []),
-            ]
-        : TBase
-      : TTopics extends readonly []
-      ? TData extends '0x'
-        ? TBase
-        : TData extends Hex
-        ? Head extends AbiParameter
-          ? Head extends { indexed: true }
-            ? Tail extends readonly AbiParameter[]
-              ? AbiEventTopicsToPrimitiveTypes<Tail, [], TData>
-              : TBase
-            : Head extends { name: infer Name extends string }
-            ? {
-                [_ in Name]: AbiParameterToPrimitiveType<Head>
-              } & (Tail extends readonly []
-                ? {}
-                : Tail extends readonly AbiParameter[]
-                ? AbiEventTopicsToPrimitiveTypes<Tail, [], TData>
-                : {})
-            : [
-                AbiParameterToPrimitiveType<Head>,
-                ...(Tail extends readonly []
-                  ? []
-                  : Tail extends readonly AbiParameter[]
-                  ? AbiEventTopicsToPrimitiveTypes<Tail, [], TData>
-                  : []),
-              ]
-          : TBase
-        : TBase
-      : TBase
-    : undefined
->