+++ src/utils/abi/decodeEventLog.test-d.ts
@@ -60,9 +60,9 @@ test('named', async () => {
 
   expectTypeOf(event).toEqualTypeOf<{
     args: {
-      from: Address
-      to: Address
-      tokenId: bigint
+      from?: Address
+      to?: Address
+      tokenId?: bigint
     }
     eventName: 'Transfer'
   }>()
@@ -100,7 +100,11 @@ test('unnamed', async () => {
   })
 
   expectTypeOf(event).toEqualTypeOf<{
-    args: [Address, Address, bigint]
+    args:
+      | readonly []
+      | readonly [`0x${string}`]
+      | readonly [`0x${string}`, `0x${string}`]
+      | readonly [`0x${string}`, `0x${string}`, bigint]
     eventName: 'Transfer'
   }>()
 })
@@ -162,17 +166,17 @@ test('unknown eventName', async () => {
   expectTypeOf(event).toEqualTypeOf<
     | {
         args: {
-          from: Address
-          to: Address
-          tokenId: bigint
+          from?: Address
+          to?: Address
+          tokenId?: bigint
         }
         eventName: 'Transfer'
       }
     | {
         args: {
-          from: Address
-          to: Address
-          value: bigint
+          from?: Address
+          to?: Address
+          value?: bigint
         }
         eventName: 'Foo'
       }