+++ src/utils/signature/recoverPublicKey.ts
@@ -18,7 +18,8 @@ export async function recoverPublicKey({
 
   // Derive v = recoveryId + 27 from end of the signature (27 is added when signing the message)
   // The recoveryId represents the y-coordinate on the secp256k1 elliptic curve and can have a value [0, 1].
-  const v = hexToNumber(`0x${signatureHex.slice(130)}`)
+  let v = hexToNumber(`0x${signatureHex.slice(130)}`)
+  if (v === 0 || v === 1) v += 27
 
   const { secp256k1 } = await import('@noble/curves/secp256k1')
   const publicKey = secp256k1.Signature.fromCompact(